added landing page for login
