+++ src/App.sass
@@ -4,8 +4,11 @@ $modal-card-body-padding: 30px;
 $control-padding-horizontal: 0.1em;
 
 @import "bulma/sass/utilities/_all.sass";
+
 @import "../node_modules/bulma/sass/elements/other.sass";
 @import "../node_modules/bulma/sass/elements/box.sass";
 @import "../node_modules/bulma/sass/elements/button.sass";
 @import "../node_modules/bulma/sass/components/modal.sass";
 @import "../node_modules/bulma/sass/form/_all.sass";
+@import "../node_modules/bulma/sass/elements/icon"
+@import "../node_modules/bulma/sass/layout/_all"
