+++ src/index.js
@@ -1,10 +1,27 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import './index.css';
-import App from './App';
-import * as serviceWorker from './serviceWorker';
+import React from "react";
+import ReactDOM from "react-dom";
+import "./index.css";
+import App from "./App";
+import LandingPage from "./pages/LandingPage";
+import * as serviceWorker from "./serviceWorker";
 
-ReactDOM.render(<App />, document.getElementById('root'));
+import { BrowserRouter, Route, Switch } from "react-router-dom";
+
+export default function Router() {
+  return (
+    <Switch>
+      <Route exact path="/" component={LandingPage} />
+      <Route exact path="/app" component={App} />
+    </Switch>
+  );
+}
+
+ReactDOM.render(
+  <BrowserRouter>
+    <Router />
+  </BrowserRouter>,
+  document.getElementById("root")
+);
 
 // If you want your app to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.
