+++ src/pages/LandingPage.js
@@ -1,0 +1,31 @@
+import React from "react";
+
+function LandingPage(props) {
+  return (
+    <div className="container">
+      <div className="field">
+        <p className="control has-icons-left has-icons-right">
+          <input className="input" type="email" placeholder="Email" />
+          <span className="icon is-small is-left">
+            <i className="fas fa-envelope"></i>
+          </span>
+          <span className="icon is-small is-right">
+            <i className="fas fa-check"></i>
+          </span>
+        </p>
+      </div>
+      <div className="field">
+        <p className="control">
+          <input className="input" type="password" placeholder="Password" />
+        </p>
+      </div>
+      <div className="field">
+        <p className="control">
+          <button className="button is-success">Login</button>
+        </p>
+      </div>
+    </div>
+  );
+}
+
+export default LandingPage;
